back/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
python-whois==0.8.0
pydantic==2.5.0
joblib==1.3.2
requests==2.31.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2